DR-354 Refactor lambda code
