+++ application-utils/common/utilities.py
@@ -1,6 +1,5 @@
 import os
 
-
 def get_table_name(table_name):
     workspace_table_name = table_name
     if os.getenv("WORKSPACE") is not None and os.getenv("WORKSPACE") != "":
